+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 - Support automated testing on windows with Travis CI (#372)
 - Support SECRETS_VERBOSE env var to enable verbosity (#323)
 - Use gpg without --quiet when decrypting in verbose mode (#394)
-- Add -v 'verbose' option to 'tell', showing gpg output (#320)
+- Add -v options to 'tell' and 'reveal', showing gpg output (#320, #395)
 - Change 'init' to never ignore .secret files (#362)
 - 'add' appends filepaths to .gitignore by default (#225)
 - Automate the GitHub release (#411)
@@ -29,6 +29,7 @@
 - Use SECRETS_GPG_COMMAND env var in gpg version check (#389)
 - Add header to git-secret.7 man page, for debian and doc improvement (#386)
 - Respect DESTDIR when installing as per GNU/debian/etc recommendations (#424)
+- Use git check-ignore to test for files ignored by git.
 
 ## Misc